fix: bug with the display of the previous image
Fixing a bug related to the display of the previous open image.
Solution: when closing the popup, delete the fields with information about the previous image (placed in the '_resetInfo' method).

There is also a solution via '.onload' images. Solution: open the popup only when the image is loaded. Lines - 23-27